requirements.txt
grpcio
tensorflow
tensorflow-serving-api
requests
flask
